move the csv util to the preview folder
